% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_differential_expression.R
\name{drop_common_suffix}
\alias{drop_common_suffix}
\title{Drop common suffix from file names
Useful to infer the data set names from differential expression files}
\usage{
drop_common_suffix(files)
}
\arguments{
\item{files}{A list of file names to process}
}
\value{
A vector with the same length as the input files containing the name of the file without
the common suffix.
}
\description{
Drop common suffix from file names
Useful to infer the data set names from differential expression files
}
