% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all.R
\name{run_all}
\alias{run_all}
\title{This runs all of the plotting functions in the package with the same configured paths.
This is used internally to generate all figures, but is not a public method
because it hard-codes the options.}
\usage{
run_all()
}
\description{
This runs all of the plotting functions in the package with the same configured paths.
This is used internally to generate all figures, but is not a public method
because it hard-codes the options.
}
