% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{save_plot_svg}
\alias{save_plot_svg}
\title{Save a ggplot to an SVG file in the manuscript font}
\usage{
save_plot_svg(plot, out_file, out_dir = ".", width = 14, height = 7)
}
\arguments{
\item{plot}{A ggplot object.}

\item{out_file}{Output file name.}

\item{out_dir}{Output directory.}

\item{width}{Plot width in inches.}

\item{height}{Plot height in inches.}
}
\value{
Invisibly, the full output path.
}
\description{
Save a ggplot to an SVG file in the manuscript font
}
