% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bican.mccarroll.figures-package.R
\docType{package}
\name{bican.mccarroll.figures-package}
\alias{bican.mccarroll.figures}
\alias{bican.mccarroll.figures-package}
\title{bican.mccarroll.differentialexpression}
\description{
Tools for differential expression and visualization of single-cell RNA-seq data
for BICAN. Provides functions for running differential expression using edgeR
and variancePartition, and visualization using GlimmaV2.
}
\author{
\strong{Maintainer}: James Nemesh \email{nemesh@broadinstitute.org}

}
\keyword{internal}
