% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bican.mccarroll.figures-package.R
\docType{package}
\name{bican.mccarroll.figures-package}
\alias{bican.mccarroll.figures}
\alias{bican.mccarroll.figures-package}
\title{bican.mccarroll.figures}
\description{
Functions that use analysis pacakges and render figures for publication.
Each function is capable of reproducing the analysis from raw data, and in long-running cases
will cache results.
}
\author{
\strong{Maintainer}: James Nemesh \email{nemesh@broadinstitute.org}

}
\keyword{internal}
